requirements.txt
attrs==21.4.0
black==22.3.0
click==8.1.3
flake8==4.0.1
iniconfig==1.1.1
joblib==1.1.0
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.23.0
packaging==21.3
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
pytest==7.1.2
scikit-learn==1.1.1
scipy==1.8.1
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.64.0
typing_extensions==4.2.0
